+++ tests/Extractors/RawPixelEncoderTest.php
@@ -3,10 +3,10 @@
 namespace Rubix\Tests\Extractors;
 
 use Rubix\ML\Extractors\Extractor;
-use Rubix\ML\Extractors\PixelEncoder;
+use Rubix\ML\Extractors\RawPixelEncoder;
 use PHPUnit\Framework\TestCase;
 
-class PixelEncoderTest extends TestCase
+class RawPixelEncoderTest extends TestCase
 {
     protected $extractor;
 
@@ -18,12 +18,12 @@ class PixelEncoderTest extends TestCase
             imagecreatefromjpeg(__DIR__ . '/space.jpg'),
         ];
 
-        $this->extractor = new PixelEncoder([3, 3], true, 0, 'gd');
+        $this->extractor = new RawPixelEncoder([3, 3], true, 'gd');
     }
 
     public function test_build_count_vectorizer()
     {
-        $this->assertInstanceOf(PixelEncoder::class, $this->extractor);
+        $this->assertInstanceOf(RawPixelEncoder::class, $this->extractor);
         $this->assertInstanceOf(Extractor::class, $this->extractor);
     }